Add a version macro to hyperv.h
Many of these modules don't explicitly set a version. So their
version ends up being an automatically generated hash that is a string
of mixed uppercase letters and numbers. The dkms script won't install
a module that does not have a higher version than an existing module
that is installed.

We can use versions like tnsr-19.05 for the module version. Dkms is
doing a check with '>' in a bash script. Lowercase t has a higher ordinal
value than any digit or uppercase letter.

We can't follow release versions exactly because 19.02-1 will evaluate
as greater than 19.02.1-1. So we omit the part after the hyphen. Probably
good enough.