remake initrd after install
On kernel upgrade, the modules were being rebuilt, but
the initramfs was being rebuilt by the kernel package
postinstall before this happened. So the system would boot
and load the old hv_vmbus module. The other modules that
depend on hv_vmbus would then fail to load.

Configure dkms to rebuild the initramfs. It's going to
happen twice now when the kernel is upgraded. Once in kernel
postinstall, once in uio_hv_generic postinstall.